+++ apps/web/core/components/pages/teams/team/team-members.tsx
@@ -15,9 +15,9 @@ import TeamMembersBlockView from './team-members-views/team-members-block-view';
 import { TOrganizationTeamEmployee } from '@/core/types/schemas';
 import { TaskCardProps } from '@/core/types/interfaces/task/task-card';
 import { useProcessedTeamMembers, useFilteredTeamMembers } from '@/core/hooks/teams/use-processed-team-members';
-import { TeamMemberFilterType } from '@/core/utils/team-members.utils';
 import { activeTeamState } from '@/core/stores';
 import { useUserQuery } from '@/core/hooks/queries/user-user.query';
+import { TeamMemberFilterType } from '@/core/lib/utils/team-members.utils';
 
 // Types for better performance and security
