+++ apps/web/core/components/tasks/task-labels.tsx
@@ -11,9 +11,9 @@ import { TaskLabelForm } from './task-labels-form';
 import { EverCard } from '../common/ever-card';
 import { Nullable } from '@/core/types/generics/utils';
 import { TTask } from '@/core/types/schemas/task/task.schema';
-import { taskUpdateQueue } from '@/core/utils/task.utils';
 import { taskLabelsListState } from '@/core/stores';
 import { useAtomValue } from 'jotai';
+import { taskUpdateQueue } from '@/core/lib/utils/task.utils';
 
 type Props = {
 	task: Nullable<TTask>;
