+++ apps/web/core/hooks/tasks/use-task-statistics.ts
@@ -17,8 +17,8 @@ import { statisticsService } from '@/core/services/client/api/timesheets/statist
 import { useRefreshIntervalV2 } from '../common';
 import { Nullable } from '@/core/types/generics/utils';
 import { TTask } from '@/core/types/schemas/task/task.schema';
-import { ITasksStatistics } from '@/core/types/interfaces/task/task';
 import { useUserQuery } from '../queries/user-user.query';
+import { TTaskStatistic } from '@/core/types/schemas/activities/statistics.schema';
 
 export function useTaskStatistics(addSeconds = 0) {
 	const { data: user } = useUserQuery();
